agent/requirements.txt
langgraph>=0.2.0
langchain-core>=0.3.0
boto3>=1.34.0
psycopg2-binary>=2.9.9
pydantic>=2.0.0
pyyaml>=6.0.0
